Move content of readme for WordPrefixPairProximityDocids into the code
